requirements.txt
fastapi
pygame
requests
uvicorn
RPi.GPIO
